requirements.txt
tencentcloud-sdk-python==3.0.668
httpx==0.23.0
click==8.1.3